refactor: 패키지 구조 변경
- MVC 패턴을 위해 패키지 구조 변경 (domain, view 추가)
- 도메인(게임, 자동차 로직 관련) : racingcar.domain
- 뷰(입출력 관련) : racingcar.view
