+++ src/main/java/racingcar/domain/GameInfo.java
@@ -1,4 +1,7 @@
-package racingcar;
+package racingcar.domain;
+
+import racingcar.domain.Car;
+import racingcar.domain.CarInfo;
 
 import java.util.List;
